Split IR graph logic into modules
Factor the build graph structures into src/ir/graph.rs and move the manifest conversion helpers into src/ir/from_manifest.rs so each Rust file stays under the 400 line limit while preserving the existing API.
