+++ src/ir.rs
@@ -23,440 +23,10 @@
 //! graph.actions.insert("a".into(), action);
 //! graph.default_targets.push(Utf8PathBuf::from("hello"));
 //! ```
-//
-use camino::Utf8PathBuf;
-use serde::Serialize;
-use std::collections::HashMap;
-use std::sync::Arc;
-use thiserror::Error;
-
-use crate::ast::{NetsukeManifest, Recipe, Rule, StringOrList};
-use crate::hasher::ActionHasher;
 
 mod cmd_interpolate;
 mod cycle;
+mod from_manifest;
+mod graph;
 
-use self::{cmd_interpolate::interpolate_command, cycle::CycleDetectionReport};
-
-/// The complete, static build graph.
-#[derive(Debug, Default, Clone)]
-pub struct BuildGraph {
-    /// All unique actions in the build keyed by a stable hash.
-    pub actions: HashMap<String, Action>,
-    /// All target files to be built keyed by output path.
-    pub targets: HashMap<Utf8PathBuf, BuildEdge>,
-    /// Targets built when no explicit target is requested.
-    pub default_targets: Vec<Utf8PathBuf>,
-}
-
-/// A reusable command analogous to a Ninja rule.
-
-#[derive(Debug, Clone, PartialEq, Serialize)]
-pub struct Action {
-    pub recipe: Recipe,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub depfile: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub deps_format: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pool: Option<String>,
-    pub restat: bool,
-}
-
-/// A single build statement connecting inputs to outputs.
-#[derive(Debug, Clone, PartialEq)]
-pub struct BuildEdge {
-    /// Identifier of the [`Action`] used for this edge.
-    pub action_id: String,
-    /// Explicit inputs that trigger a rebuild when changed.
-    pub inputs: Vec<Utf8PathBuf>,
-    /// Outputs explicitly generated by the command.
-    pub explicit_outputs: Vec<Utf8PathBuf>,
-    /// Outputs implicitly generated by the command (Ninja `|`).
-    pub implicit_outputs: Vec<Utf8PathBuf>,
-    /// Order-only dependencies that do not trigger rebuilds (Ninja `||`).
-    pub order_only_deps: Vec<Utf8PathBuf>,
-    /// Output does not correspond to a real file.
-    pub phony: bool,
-    /// Run the command on every invocation regardless of timestamps.
-    pub always: bool,
-}
-
-/// Errors produced during IR generation.
-///
-/// Each variant documents a distinct validation failure encountered while
-/// constructing the intermediate representation from a manifest.
-///
-/// # Examples
-///
-/// ```
-/// use netsuke::ir::IrGenError;
-///
-/// fn describe(err: IrGenError) -> String {
-///     match err {
-///         IrGenError::EmptyRule { target_name } => {
-///             format!("{target_name} missing rule")
-///         },
-///         other => other.to_string(),
-///     }
-/// }
-///
-/// assert_eq!(
-///     describe(IrGenError::EmptyRule {
-///         target_name: "app".into(),
-///     }),
-///     "app missing rule"
-/// );
-/// ```
-#[derive(Debug, Error)]
-pub enum IrGenError {
-    /// Raised when a target references a rule that is not defined in the
-    /// manifest.
-    ///
-    /// ```
-    /// use netsuke::ir::IrGenError;
-    ///
-    /// let err = IrGenError::RuleNotFound {
-    ///     target_name: "app".into(),
-    ///     rule_name: "compile".into(),
-    /// };
-    /// assert!(matches!(
-    ///     err,
-    ///     IrGenError::RuleNotFound { rule_name, .. }
-    ///         if rule_name == "compile"
-    /// ));
-    /// ```
-    #[error("rule '{rule_name}' referenced by target '{target_name}' was not found")]
-    RuleNotFound {
-        target_name: String,
-        rule_name: String,
-    },
-
-    /// Triggered when multiple rule names are supplied for a single target.
-    ///
-    /// ```
-    /// use netsuke::ir::IrGenError;
-    ///
-    /// let err = IrGenError::MultipleRules {
-    ///     target_name: "lib".into(),
-    ///     rules: vec!["c".into(), "cpp".into()],
-    /// };
-    /// if let IrGenError::MultipleRules { rules, .. } = err {
-    ///     assert_eq!(
-    ///         rules,
-    ///         vec!["c".to_string(), "cpp".to_string()]
-    ///     );
-    /// }
-    /// ```
-    #[error("multiple rules for target '{target_name}': {rules:?}")]
-    MultipleRules {
-        target_name: String,
-        rules: Vec<String>,
-    },
-
-    /// Returned when a target declares no rule at all.
-    ///
-    /// ```
-    /// use netsuke::ir::IrGenError;
-    ///
-    /// let err = IrGenError::EmptyRule { target_name: "docs".into() };
-    /// assert_eq!(
-    ///     err.to_string(),
-    ///     "No rules specified for target docs"
-    /// );
-    /// ```
-    #[error("No rules specified for target {target_name}")]
-    EmptyRule { target_name: String },
-
-    /// Indicates that more than one build edge produces the same output file.
-    ///
-    /// ```
-    /// use netsuke::ir::IrGenError;
-    ///
-    /// let err = IrGenError::DuplicateOutput {
-    ///     outputs: vec!["obj.o".into()],
-    /// };
-    /// if let IrGenError::DuplicateOutput { outputs } = err {
-    ///     assert_eq!(
-    ///         outputs,
-    ///         vec!["obj.o".to_string()]
-    ///     );
-    /// }
-    /// ```
-    #[error("duplicate target outputs: {outputs:?}")]
-    DuplicateOutput { outputs: Vec<String> },
-
-    /// Emitted when a cycle exists in the target graph.
-    ///
-    /// ```
-    /// use camino::Utf8PathBuf;
-    /// use netsuke::ir::IrGenError;
-    ///
-    /// let err = IrGenError::CircularDependency {
-    ///     cycle: vec![Utf8PathBuf::from("a"), Utf8PathBuf::from("a")],
-    ///     missing_dependencies: Vec::new(),
-    /// };
-    /// if let IrGenError::CircularDependency { cycle, .. } = err {
-    ///     assert_eq!(
-    ///         cycle,
-    ///         vec![Utf8PathBuf::from("a"), Utf8PathBuf::from("a")]
-    ///     );
-    /// }
-    /// ```
-    #[error("circular dependency detected: {cycle:?}")]
-    CircularDependency {
-        cycle: Vec<Utf8PathBuf>,
-        missing_dependencies: Vec<(Utf8PathBuf, Utf8PathBuf)>,
-    },
-
-    /// Wraps failures encountered while serialising an action to JSON.
-    ///
-    /// ```
-    /// use netsuke::ir::IrGenError;
-    ///
-    /// let source = serde_json::Error::custom("invalid action");
-    /// let err = IrGenError::ActionSerialisation(source);
-    /// assert!(err.to_string().contains("invalid action"));
-    /// ```
-    #[error("failed to serialise action: {0}")]
-    ActionSerialisation(#[from] serde_json::Error),
-
-    /// Raised when command interpolation yields an invalid shell snippet.
-    ///
-    /// ```
-    /// use netsuke::ir::IrGenError;
-    ///
-    /// let err = IrGenError::InvalidCommand {
-    ///     command: "echo $in".into(),
-    ///     snippet: "echo $in".into(),
-    /// };
-    /// assert_eq!(
-    ///     err.to_string(),
-    ///     "command is not a valid shell command: echo $in"
-    /// );
-    /// ```
-    #[error("command is not a valid shell command: {snippet}")]
-    InvalidCommand { command: String, snippet: String },
-}
-
-impl BuildGraph {
-    /// Transform a manifest into a [`BuildGraph`].
-    ///
-    /// # Errors
-    ///
-    /// Returns [`IrGenError`] when a referenced rule is missing, multiple rules
-    /// are specified for a single target, or no rule is provided.
-    pub fn from_manifest(manifest: &NetsukeManifest) -> Result<Self, IrGenError> {
-        let mut graph = Self::default();
-        let mut rule_map: HashMap<String, Arc<Rule>> = HashMap::new();
-
-        Self::process_rules(manifest, &mut rule_map);
-        Self::process_targets(manifest, &mut graph.actions, &mut graph.targets, &rule_map)?;
-        Self::process_defaults(manifest, &mut graph.default_targets);
-
-        graph.detect_cycles()?;
-
-        Ok(graph)
-    }
-
-    /// Collect rule templates without deduplicating them.
-    ///
-    /// Rules are stored verbatim and expanded later when targets reference
-    /// them. This allows each target's input and output paths to be embedded in
-    /// the resulting command, meaning identical rule definitions may yield
-    /// distinct actions once interpolated. Should the manifest schema ever
-    /// permit targets to override recipe fields such as `command` or
-    /// `description`, those target-level values take precedence over the rule's
-    /// defaults.
-    fn process_rules(manifest: &NetsukeManifest, rule_map: &mut HashMap<String, Arc<Rule>>) {
-        for rule in &manifest.rules {
-            rule_map.insert(rule.name.clone(), Arc::new(rule.clone()));
-        }
-    }
-
-    fn process_targets(
-        manifest: &NetsukeManifest,
-        actions: &mut HashMap<String, Action>,
-        targets: &mut HashMap<Utf8PathBuf, BuildEdge>,
-        rule_map: &HashMap<String, Arc<Rule>>,
-    ) -> Result<(), IrGenError> {
-        for target in manifest.actions.iter().chain(&manifest.targets) {
-            let outputs = to_paths(&target.name);
-            let inputs = to_paths(&target.sources);
-            let target_name = get_target_display_name(&outputs);
-            let action_id = match &target.recipe {
-                Recipe::Rule { rule } => {
-                    let tmpl = resolve_rule(rule, rule_map, &target_name)?;
-                    // Future schema versions may allow targets to override
-                    // recipe or description fields. If so, those values will
-                    // take precedence over the rule template.
-                    register_action(
-                        actions,
-                        tmpl.recipe.clone(),
-                        tmpl.description.clone(),
-                        &inputs,
-                        &outputs,
-                    )?
-                }
-                Recipe::Command { .. } | Recipe::Script { .. } => {
-                    register_action(actions, target.recipe.clone(), None, &inputs, &outputs)?
-                }
-            };
-
-            let edge = BuildEdge {
-                action_id,
-                inputs: inputs.clone(),
-                explicit_outputs: outputs.clone(),
-                implicit_outputs: Vec::new(),
-                order_only_deps: to_paths(&target.order_only_deps),
-                phony: target.phony,
-                always: target.always,
-            };
-
-            if let Some(dups) = find_duplicates(&outputs, targets) {
-                return Err(IrGenError::DuplicateOutput { outputs: dups });
-            }
-            for out in outputs {
-                targets.insert(out, edge.clone());
-            }
-        }
-        Ok(())
-    }
-
-    fn process_defaults(manifest: &NetsukeManifest, defaults: &mut Vec<Utf8PathBuf>) {
-        defaults.extend(manifest.defaults.iter().map(Utf8PathBuf::from));
-    }
-
-    fn detect_cycles(&self) -> Result<(), IrGenError> {
-        let CycleDetectionReport {
-            cycle,
-            missing_dependencies,
-        } = cycle::analyse(&self.targets);
-        if let Some(cycle) = cycle {
-            return Err(IrGenError::CircularDependency {
-                cycle,
-                missing_dependencies,
-            });
-        }
-        Ok(())
-    }
-}
-
-/// Insert an action into the graph, deduplicating on the resolved command and
-/// file set.
-///
-/// The rule template is interpolated with the target's inputs and outputs,
-/// which are shell-escaped and embedded directly in the command. The resulting
-/// [`Action`] is hashed so identical commands operating on the same file sets
-/// share an identifier, while differing paths produce distinct actions even if
-/// they originate from the same rule definition.
-fn register_action(
-    actions: &mut HashMap<String, Action>,
-    recipe: Recipe,
-    description: Option<String>,
-    inputs: &[Utf8PathBuf],
-    outputs: &[Utf8PathBuf],
-) -> Result<String, IrGenError> {
-    let recipe = match recipe {
-        Recipe::Command { command } => {
-            let interpolated = interpolate_command(&command, inputs, outputs)?;
-            Recipe::Command {
-                command: interpolated,
-            }
-        }
-        other => other,
-    };
-    let action = Action {
-        recipe,
-        description,
-        depfile: None,
-        deps_format: None,
-        pool: None,
-        restat: false,
-    };
-    let hash = ActionHasher::hash(&action).map_err(IrGenError::ActionSerialisation)?;
-    actions.entry(hash.clone()).or_insert(action);
-    Ok(hash)
-}
-
-fn map_string_or_list<T, F>(sol: &StringOrList, f: F) -> Vec<T>
-where
-    F: Fn(&str) -> T,
-{
-    match sol {
-        StringOrList::Empty => Vec::new(),
-        StringOrList::String(s) => vec![f(s)],
-        StringOrList::List(v) => v.iter().map(|s| f(s)).collect(),
-    }
-}
-
-fn to_paths(sol: &StringOrList) -> Vec<Utf8PathBuf> {
-    map_string_or_list(sol, |s| Utf8PathBuf::from(s))
-}
-
-fn to_string_vec(sol: &StringOrList) -> Vec<String> {
-    map_string_or_list(sol, str::to_string)
-}
-
-fn extract_single(sol: &StringOrList) -> Option<&str> {
-    match sol {
-        StringOrList::String(s) => Some(s),
-        StringOrList::List(v) if v.len() == 1 => v.first().map(String::as_str),
-        _ => None,
-    }
-}
-
-fn resolve_rule(
-    rule: &StringOrList,
-    rule_map: &HashMap<String, Arc<Rule>>,
-    target_name: &str,
-) -> Result<Arc<Rule>, IrGenError> {
-    extract_single(rule).map_or_else(
-        || {
-            let mut rules = to_string_vec(rule);
-            if rules.is_empty() {
-                Err(IrGenError::EmptyRule {
-                    target_name: target_name.to_string(),
-                })
-            } else {
-                rules.sort();
-                Err(IrGenError::MultipleRules {
-                    target_name: target_name.to_string(),
-                    rules,
-                })
-            }
-        },
-        |name| {
-            rule_map
-                .get(name)
-                .cloned()
-                .ok_or_else(|| IrGenError::RuleNotFound {
-                    target_name: target_name.to_string(),
-                    rule_name: name.to_string(),
-                })
-        },
-    )
-}
-
-fn find_duplicates(
-    outputs: &[Utf8PathBuf],
-    targets: &HashMap<Utf8PathBuf, BuildEdge>,
-) -> Option<Vec<String>> {
-    let mut dups: Vec<String> = outputs
-        .iter()
-        .filter(|o| targets.contains_key(*o))
-        .map(|p| p.as_str().to_owned())
-        .collect();
-    if dups.is_empty() {
-        None
-    } else {
-        dups.sort();
-        Some(dups)
-    }
-}
-
-fn get_target_display_name(paths: &[Utf8PathBuf]) -> String {
-    paths.first().map(ToString::to_string).unwrap_or_default()
-}
+pub use graph::{Action, BuildEdge, BuildGraph, IrGenError};
